[Radoub] refactor: Extract test helpers to FlaUITestBase (#693)
Deduplicated FindElement, FindTabByName, FindMenu, FindPopupByTitle,
and IsElementVisible methods from 7 test files to shared base class.
